Swapped SBCN's SF/EB fields
The SF used to be 3 bits and EB 4bits.  This meant the SF of 7 resulted in
a max Sframe of 128 Tslots (32s)... not enough for future expansion.
Swapping SF/EB in the frame_spec allows SF Order to be 15 (4096 Tslots, 17 minutes).  Future work will be to have EB become a *multiple* of SF