Moved phy-config from MAC layer to PHY init
There are two kinds of config: default config and runtime config.
The conents of phy_cfg.py is the default config.
This is what was moved to the PHY's initializing state handler.

A couple PHY config items remain in MAC at this moment: the TX and RX freqs.
These are run-time config items because we remain open to frequency hopping as a future possibility.
I am looking for a way to pull run-time config items out of phy_cfg.py to clarify their different role.