Fix NMEA parsing (buffer overgrowth)
Problem was NMEA parsing function was creating a slice
of the bytearray it was passed (and meant to modify and return by reference).
The slice was being modified and not the original bytearray.
Fix was to keep a reference to the original bytearray
and set its contents to the modified bytearray at the end of the function.

Fixes issue #2