Improved the new-user experience
Added an example .env path for Windows.
Fixed an uncaught exception in the mock serial module.
Added an assert to tell the user to create a missing credential file.
Added comments on how to use pip to install specific versions of modules.
Added a Getting Started section to the README