Filling in the PHY layer public interface
0) Added a class initializer parameter to command the state machine
to either listen or sleep by default.
1) Added a method to start the stack from the bottom up
2) Added a method to enqueue a frame for transmission
A frame of data (bytes object), a time to transmit and tx settings
are given from the LNK layer to the PHY.
A dual-queue system is implemented: one queue for immediate tx
another queue (dict actually) for timed tx.
3) Added a _RESCHEDULE signal and event that is used to break
the state machine out of its normal mode (rx or sleep)
when new data is enqueued for tx.

Also changed old "mac" things to "phy"

Status:
The PHY layer is partially operational.
When run with an example program,
I see a transmission is happening (green led flash).
But since settings are not being applied,
who knows what freq it's on.