Fixed and promoted  frame.is_* methods
The frame.is_* methods check the bits of the Fctl field
for a specific property.  The problem was that the bit-test
was assuming a non-extended type frame.  The fix is to
add an explicit test that the frame is not extended.

The promotion is from private to public scope;
something needed for upcoming work.

The frame.is_* methods are in order of the Fctl bitfield
rather than alpha order.