Remove _DIO_PAYLD_CRC_ERR handler
The SX127x generates the RxDone and PayldCrcErr signals at the same time.
When a PayldCrcErr is generated, the fifo is modified and fifo ptrs
must be reset.
So it makes no sense to respond to the PayldCrcErr pin signal.
Removing this handler lets the RxDone handler evaluate
PayldCrcErr irq flag and process the fifo if needed.

In short, removing dead code.

Fixes #6