Do not use DIO5/ModeReady
Not many OEM LoRa modules offer DIO5 as an option.
So lets not require its use in this project.
Previously, ModeReady was only used in one location:
after a software timeout during transmit when the modem
opmode was set to standby.
Now, when entering _scheduling state,
if the opmode is not Standby a 100ms delay is issued.

A side effect of this change is that the write_opmode()
signature is able to lose the second argument which
optionally enabled/disabled the DIO5/ModeReady interrupt.

Fixes #5