Fix _txing@_PHY_TMOUT trouble
Problem was: when TXing rapid back-to-back messages,
sometimes the TxDone signals would stop and the software
timeout timer would be needed to exit the _txing() state.

Normal path: if the messages were TX'd slowly, the state machine
would transition through _scheduling and _listening which would
have a side effect of clearing the interrupt flags.
Error path: When the messages are TX'd rapidly,  the transitions
do NOT go through _listening.  They go through _scheduling and
_txing and the TxDone interrupt flag was NOT being cleared.

Solution: is to clear the TxDone interrupt flag in the _DIO_TX_DONE handler.

Fixes #7