Refactor HeymacCommands
The problem was that I was writing different code to serialize
and parse each cmd class.  The fix was to:
1) impose a format limitation on commands
to make serialization and parsing easier.
The limitation is that variable-length commands must be
a variable number of fixed-format entries.
An entry is a sequence of data that can be [un]packed by
Python's built-in struct module.
2) Perform serialization and parsing in the base class.
3) Define only the unique data (the entry format) in cmd subclasses.

The other clever refactor here is that each class defines a namedtuple
for its entry type.  This way, the output of struct.unpack() can be
fed straight into the namedtuple constructor.  This lets the entry's
data have named fields.  Then it was simply a matter of adding
a __getattr__() to make accessing cmd instance field as named attribute.
Those same names can be used as keyword args to the constructor.
Very pythonic!